+++ src/test/java/ru/apolonov/helpers/AllureAttachments.java
@@ -1,57 +1,57 @@
 package ru.apolonov.helpers;
 
-import io.qameta.allure.Allure;
 import io.qameta.allure.Attachment;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.remote.RemoteWebDriver;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
+import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 
-import static com.codeborne.selenide.Selenide.sleep;
+import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
+import static ru.apolonov.helpers.DriverUtils.getConsoleLogs;
 
 public class AllureAttachments {
-    public static final Logger LOGGER = LoggerFactory.getLogger(AllureAttachments.class);
 
     @Attachment(value = "{attachName}", type = "text/plain")
-    private static String addMessage(String attachName, String text) {
-        return text;
+    public static String attachAsText(String attachName, String message) {
+        return message;
     }
 
-    public static void addBrowserConsoleLogs() {
-        addMessage("Browser console logs", DriverUtils.getConsoleLogs());
+    @Attachment(value = "Page source", type = "text/plain")
+    public static byte[] pageSource() {
+        return getWebDriver().getPageSource().getBytes(StandardCharsets.UTF_8);
     }
 
     @Attachment(value = "{attachName}", type = "image/png")
-    public static byte[] addScreenshotAs(String attachName) {
-        return DriverUtils.getScreenshotAsBytes();
+    public static byte[] screenshotAs(String attachName) {
+        return ((TakesScreenshot) getWebDriver()).getScreenshotAs(OutputType.BYTES);
     }
 
-    @Attachment(value = "Page source", type = "text/html")
-    public static byte[] addPageSource() {
-        return DriverUtils.getPageSourceAsBytes();
+    public static void browserConsoleLogs() {
+        attachAsText("Browser console logs", getConsoleLogs());
     }
 
-    public static void addVideo(String sessionId) {
-        URL videoUrl = DriverUtils.getVideoUrl(sessionId);
-        if (videoUrl != null) {
-            InputStream videoInputStream = null;
-            sleep(1000);
+    @Attachment(value = "Video", type = "text/html", fileExtension = ".html")
+    public static String addVideo() {
+        return "<html><body><video width='100%' height='100%' controls autoplay><source src='"
+                + getVideoUrl(getSessionId())
+                + "' type='video/mp4'></video></body></html>";
+    }
 
-            for (int i = 0; i < 10; i++) {
-                try {
-                    videoInputStream = videoUrl.openStream();
-                    break;
-                } catch (FileNotFoundException e) {
-                    sleep(1000);
-                } catch (IOException e) {
-                    LOGGER.warn("[ALLURE VIDEO ATTACHMENT ERROR] Cant attach allure video, {}", videoUrl);
-                    e.printStackTrace();
-                }
-            }
-            Allure.addAttachment("Video", "video/mp4", videoInputStream, "mp4");
+    public static URL getVideoUrl(String sessionId) {
+        String videoUrl = "https://selenoid.autotests.cloud/video/" + sessionId + ".mp4";
+
+        try {
+            return new URL(videoUrl);
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
         }
+        return null;
+    }
+
+    public static String getSessionId() {
+        return ((RemoteWebDriver) getWebDriver()).getSessionId().toString();
     }
 }
