+++ src/test/java/ru/apolonov/helpers/DriverSettings.java
@@ -25,11 +25,13 @@ public class DriverSettings {
         chromeOptions.addArguments("--disable-notifications");
         chromeOptions.addArguments("--lang=en-en");
 
-        if (Project.isWebMobile()) { // for chrome only
+       /* if (Project.isWebMobile()) { // for chrome only
             Map<String, Object> mobileDevice = new HashMap<>();
             mobileDevice.put("deviceName", Project.config.browserMobileView());
             chromeOptions.setExperimentalOption("mobileEmulation", mobileDevice);
         }
+
+        */
 
         if (Project.isRemoteWebDriver()) {
             capabilities.setCapability("enableVNC", true);
@@ -39,5 +41,6 @@ public class DriverSettings {
 
         capabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
         Configuration.browserCapabilities = capabilities;
+        Configuration.timeout = 10000;
     }
 }
